+++ eras/babbage/impl/CHANGELOG.md
@@ -1,7 +1,15 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.13.1.0
+## 1.14.0.0
 
+* Deprecate the use of `GetLedgerView`:
+  - Add `BabbageForecast` to deprecate `LedgerView` for Praos.
+    + `mkBabbageForecast`
+    + `bfPoolDistrL`
+    + `bfMaxBlockHeaderSizeL`
+    + `bfMaxBlockBodySizeL`
+    + `bfProtocolVersionL`
+  - Add `EraForecast` instance with `BabbageForecast`.
 * Add `validateScriptsWellFormedTxOuts`
 
 ## 1.13.0.0